modularize: configuration and structures
