+++ src/data/DataStructures.cpp
@@ -1,0 +1,6 @@
+#include "DataStructures.h"
+
+// Global data structures
+PowerData PhasePower[3];
+EnergyData PhaseEnergy[3];
+String serJsonResponse;
